requirements.txt
anyio==3.6.1
arel==0.2.0
asgiref==3.5.2
bidict==0.22.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
fastapi==0.74.0
h11==0.13.0
httptools==0.3.0
idna==3.3
imageio==2.19.5
imgaug==0.4.0
install==1.3.5
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.3
numpy==1.23.3
opencv-python==4.6.0.66
pandas==1.4.2
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
scikit-image==0.19.3
scikit-video==1.1.11
starlette==0.17.1
tdqm==0.0.1
tqdm==4.64.1
uvicorn==0.17.0
varname==0.8.3
wincertstore==0.2
